requirements.txt
PyQt5
google.generativeai
python-dotenv